Move AbortSignal into its own file
