+++ types/node-fetch/tsconfig.json
@@ -19,6 +19,7 @@
     },
     "files": [
         "index.d.ts",
-        "node-fetch-tests.ts"
+        "node-fetch-tests.ts",
+        "AbortSignal.d.ts"
     ]
 }
